+++ projects/vue2-libs/components/v-perfect-scrollbar/classes/SCROLL_DIRECTION.enum.js
@@ -1,0 +1,9 @@
+/** 滚动方向
+ *
+ *
+ * @export
+ */
+export const SCROLL_DIRECTION = {
+    SCROLL_TOP: 'scrollTop',
+    SCROLL_LEFT: 'scrollLeft',
+};