+++ projects/vue2-libs/components/v-perfect-scrollbar/index.js
@@ -1,0 +1,10 @@
+import VPerfectScrollbar from './v-perfect-scrollbar.vue';
+export {
+    VPerfectScrollbar
+};
+export {
+    SCROLL_DIRECTION
+} from './classes';
+export {
+    VPerfectScrollbarService
+} from './v-perfect-scrollbar.service';