+++ projects/vue2-libs/services/index.js
@@ -1,0 +1,1 @@
+export { EventListenService } from './event-listen.service';