+++ projects/vue2-libs/changelog.md
@@ -1,4 +1,9 @@
 # V2.0.0
+## 2019-03-10
+### vue2-libs
+- v-perfect-scrollbar:增加perfect-scrollbar滚动条组件（wangjx）；
+- event-listen.service:增加event-listen事件监听服务（wangjx）；
+
 ## 2019-02-17
 ### vue2-libs
 - ngx-primejs:增加primejs代码高亮组件（wangjx）；